Rename reference_system_autoware into reference_system
