+++ reference_system/include/reference_system/node_settings.hpp
