+++ reference_system/include/reference_system/reference_nodes/command.hpp
@@ -11,17 +11,17 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#ifndef REFERENCE_SYSTEM_AUTOWARE__NODE__COMMAND_HPP_
-#define REFERENCE_SYSTEM_AUTOWARE__NODE__COMMAND_HPP_
+#ifndef REFERENCE_SYSTEM__NODE__COMMAND_HPP_
+#define REFERENCE_SYSTEM__NODE__COMMAND_HPP_
 #pragma once
 
 #include <chrono>
 #include <string>
 
 #include "rclcpp/rclcpp.hpp"
-#include "reference_system_autoware/node_settings.hpp"
-#include "reference_system_autoware/sample_management.hpp"
-#include "reference_system_autoware/types.hpp"
+#include "reference_system/node_settings.hpp"
+#include "reference_system/sample_management.hpp"
+#include "reference_system/types.hpp"
 
 namespace node
 {
@@ -48,4 +48,4 @@ private:
 };
 }  // namespace node
 
-#endif  // REFERENCE_SYSTEM_AUTOWARE__NODE__COMMAND_HPP_
+#endif  // REFERENCE_SYSTEM__NODE__COMMAND_HPP_
