+++ reference_system/include/reference_system/reference_nodes/sensor.hpp
@@ -11,8 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#ifndef REFERENCE_SYSTEM_AUTOWARE__NODE__SENSOR_HPP_
-#define REFERENCE_SYSTEM_AUTOWARE__NODE__SENSOR_HPP_
+#ifndef REFERENCE_SYSTEM__NODE__SENSOR_HPP_
+#define REFERENCE_SYSTEM__NODE__SENSOR_HPP_
 #pragma once
 
 #include <chrono>
@@ -20,9 +20,9 @@
 #include <utility>
 
 #include "rclcpp/rclcpp.hpp"
-#include "reference_system_autoware/node_settings.hpp"
-#include "reference_system_autoware/sample_management.hpp"
-#include "reference_system_autoware/types.hpp"
+#include "reference_system/node_settings.hpp"
+#include "reference_system/sample_management.hpp"
+#include "reference_system/types.hpp"
 
 namespace node
 {
@@ -56,4 +56,4 @@ private:
 };
 }  // namespace node
 
-#endif  // REFERENCE_SYSTEM_AUTOWARE__NODE__SENSOR_HPP_
+#endif  // REFERENCE_SYSTEM__NODE__SENSOR_HPP_
