+++ reference_system/include/reference_system/types.hpp
@@ -11,24 +11,16 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+#ifndef REFERENCE_SYSTEM__TYPES_HPP_
+#define REFERENCE_SYSTEM__TYPES_HPP_
+#pragma once
 
 #include "rclcpp/rclcpp.hpp"
+#include "reference_interfaces/msg/message4kb.hpp"
+#include "std_msgs/msg/string.hpp"
 
-#include "reference_system_autoware/reference_system.hpp"
+using message_t = reference_interfaces::msg::Message4kb;
+using publisher_t = rclcpp::Publisher<message_t>::SharedPtr;
+using subscription_t = rclcpp::Subscription<message_t>::SharedPtr;
 
-int main(int argc, char* argv[]) {
-  rclcpp::init(argc, argv);
-
-  auto nodes = create_reference_system_nodes();
-
-  rclcpp::executors::MultiThreadedExecutor executor;
-  for (auto& node : nodes) {
-    executor.add_node(node);
-  }
-  executor.spin();
-
-  nodes.clear();
-  rclcpp::shutdown();
-
-  return 0;
-}
+#endif  // REFERENCE_SYSTEM__TYPES_HPP_
