+++ reference_system/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>reference_system_autoware</name>
+  <name>reference_system</name>
   <version>0.0.0</version>
   <description>Reference system for the ROScon workshop</description>
   <maintainer email="evan.flynn@apex.ai">Evan Flynn</maintainer>
