+++ reference_system/src/ros2/executor/autoware_default_multithreaded.cpp
@@ -1,0 +1,34 @@
+// Copyright 2021 Apex.AI, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "rclcpp/rclcpp.hpp"
+
+#include "reference_system/reference_system.hpp"
+
+int main(int argc, char* argv[]) {
+  rclcpp::init(argc, argv);
+
+  auto nodes = create_reference_system_nodes();
+
+  rclcpp::executors::MultiThreadedExecutor executor;
+  for (auto& node : nodes) {
+    executor.add_node(node);
+  }
+  executor.spin();
+
+  nodes.clear();
+  rclcpp::shutdown();
+
+  return 0;
+}
