+++ reference_system/src/ros2/executor/autoware_default_singlethreaded.cpp
@@ -11,16 +11,24 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#ifndef REFERENCE_SYSTEM_AUTOWARE__TYPES_HPP_
-#define REFERENCE_SYSTEM_AUTOWARE__TYPES_HPP_
-#pragma once
 
 #include "rclcpp/rclcpp.hpp"
-#include "reference_interfaces/msg/message4kb.hpp"
-#include "std_msgs/msg/string.hpp"
 
-using message_t = reference_interfaces::msg::Message4kb;
-using publisher_t = rclcpp::Publisher<message_t>::SharedPtr;
-using subscription_t = rclcpp::Subscription<message_t>::SharedPtr;
+#include "reference_system/reference_system.hpp"
 
-#endif  // REFERENCE_SYSTEM_AUTOWARE__TYPES_HPP_
+int main(int argc, char* argv[]) {
+  rclcpp::init(argc, argv);
+
+  auto nodes = create_reference_system_nodes();
+
+  rclcpp::executors::SingleThreadedExecutor executor;
+  for (auto& node : nodes) {
+    executor.add_node(node);
+  }
+  executor.spin();
+
+  nodes.clear();
+  rclcpp::shutdown();
+
+  return 0;
+}
