+++ reference_system/test/generate_graph.sh
